requirements.txt
# NO ML libraries — all LLM calls happen externally via inference.py
openenv-core==0.2.3
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.11.7
websockets==16.0
python-multipart==0.0.24
